login one_many_relation

+++ app/Http/Controllers/TasksController.php
@@ -11,11 +11,28 @@ class TasksController extends Controller
 
     public function index()
     {
-        $tasks = Task::all();
+        // $tasks = Task::all();
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        // ]);
+        
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            return view('tasks.index', $data);
+        } else {
+            return view('welcome');
+        }
+        
+        
     }
 
     public function create()
@@ -37,10 +54,11 @@ class TasksController extends Controller
         
         $task = new Task;
         $task->status = $request->status; //追加
+        $task->user_id = $request->user()->id; //追加
         $task->content = $request->content;
         $task->save();
         
-        return redirect('/');
+        return redirect('/tasks/list');
     }
 
     public function show($id)
@@ -70,17 +88,28 @@ class TasksController extends Controller
         
         $task = Task::find($id);
         $task->status = $request->status; //追加
+        $task->user_id = $request->user()->id; //追加
         $task->content = $request->content;
-        $task->save();
         
-        return redirect('/');
+        // $task->save();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->save();
+        }
+        
+        return redirect('/tasks/list');
     }
 
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
+        // $task->delete();
+        
+        return redirect('/tasks/list');
     }
 }
